% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{AddColBinnedToQuartiles}
\alias{AddColBinnedToQuartiles}
\title{Bin the values of a selected continuous column into 4 bins (quartiles) and add the bin assignments as a new column}
\usage{
AddColBinnedToQuartiles(inputted.data, col.name.to.bin, name.of.new.col)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{col.name.to.bin}{Name of a continuous column to use the values for binning.}

\item{name.of.new.col}{Name of the new column that will be added to the end of the dataframe. A suitable name would be "TARGET.quartile"}
}
\value{
A dataframe with bin assignment added as a factor column.
}
\description{
Bin the values of a selected continuous column into 4 bins (quartiles) and add the bin assignments as a new column
}
