% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{captureSessionInfo}
\alias{captureSessionInfo}
\title{Capture session info}
\usage{
captureSessionInfo()
}
\value{
No object is returned. A txt file is created that captures the sessionInfo() output.
}
\description{
Capture session info to keep track of what packages and what versions are
being used in the current session. Writes the session info output to a txt
file.
}
