% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CorAssoTestMultipleWithErrorHandling}
\alias{CorAssoTestMultipleWithErrorHandling}
\title{Takes multiple vectors and do correlation/association testing with all of them}
\usage{
CorAssoTestMultipleWithErrorHandling(
  data.vectors1,
  variable.types1,
  names1,
  data.vectors2,
  variable.types2,
  names2,
  table.name
)
}
\arguments{
\item{data.vectors1}{List of numerical vectors.}

\item{variable.types1}{A vector of strings where each element can be: "continuous", "categorical", or "ordinal". Specifies data type of vectors in data.vectors1.}

\item{names1}{A vector strings that specify the names that should go with each numerical vector in data.vectors1.}

\item{data.vectors2}{List of numerical vectors.}

\item{variable.types2}{A vector of strings where each element can be: "continuous", "categorical", or "ordinal". Specifies data type of vectors in data.vectors2.}

\item{names2}{A vector strings that specify the names that should go with each numerical vector in data.vectors2.}

\item{table.name}{Name of the outputted text file.}
}
\value{
No object is returned, but function creates a text file with the p-values from the testings.
}
\description{
Uses correlation.association.test() for multiple pairs of data vectors and
then outputs results into a text file.
}
\details{
This function can be used for batch effect testing. Example: We can ask
if the dependent variable significantly differs based on gender or race.
If it does, then it could impact the relationship of the independent
variable with the dependent variable. See ancovall for functions that
can adjust for covariates like gender and race.
}
