% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{correlation.association.test}
\alias{correlation.association.test}
\title{Given two numerical data vector, determine the correlation}
\usage{
correlation.association.test(
  data1,
  data1.type,
  data1.name,
  data2,
  data2.type,
  data2.name
)
}
\arguments{
\item{data1}{Numerical data vector.}

\item{data1.type}{A string to specify what type of data is data1. Options are "continuous", "ordinal", "categorical".}

\item{data1.name}{String that specifies name to associate with data1.}

\item{data2}{Numerical data vector.}

\item{data2.type}{A string to specify what type of data is data1. Options are "continuous", "ordinal", "categorical".}

\item{data2.name}{String that specifies name to associate with data1.}
}
\value{
A vector of strings with two elements:
\enumerate{
\item Description of test performed
\item P-value
}
}
\description{
This is a work in progress function.
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
