% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{describeNumericalColumns}
\alias{describeNumericalColumns}
\title{Describe each numerical feature. Mean, stddev, median, skewness (symmetry), kurtosis (flatness), pass normality?}
\usage{
describeNumericalColumns(input.data, column.names.to.use, file.name = NULL)
}
\arguments{
\item{input.data}{A dataframe.}

\item{column.names.to.use}{Vector of strings of column names. Function will only describe these specified columns (features).}

\item{file.name}{Name of the text file or absolute path of the file to write the output to. If the input is NULL, then
no file is generated.}
}
\value{
A dataframe where the rows are the descriptions and columns are the features.
Additionally, a text file will be created which also captured the function output.
}
\description{
Describe each numerical feature. Mean, stddev, median, skewness (symmetry), kurtosis (flatness), pass normality?
}
