% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{describeNumericalColumnsWithLevels}
\alias{describeNumericalColumnsWithLevels}
\title{For each level, describe each numerical feature. Mean, sd, median, skewness (symmetry), kurtosis (flatness), pass normality?}
\usage{
describeNumericalColumnsWithLevels(
  input.data,
  column.names.to.use,
  column.name.containing.levels.within.feature,
  file.name = NULL
)
}
\arguments{
\item{input.data}{A dataframe.}

\item{column.names.to.use}{Vector of strings of column names. Function will only describe these specified columns (features).}

\item{column.name.containing.levels.within.feature}{Name of column that contains the levels that you want to separate by.}

\item{file.name}{Name of the text file or absolute path of the file to write the output to. If the input is NULL, then
no file is generated.}
}
\value{
A dataframe where the rows are the descriptions and columns are the features.
Additionally, a text file will be created which also captured the function output.

The output for each level is separated by blank rows.
}
\description{
Uses describeNumericalColumns() separately for each unique level in the column
corresponding to column.name.containing.levels.within.feature. The results
for each level are separated in the output by blank rows.
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
