% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{DownSampleDataframe}
\alias{DownSampleDataframe}
\title{Down sample an imbalanced dataset to get a balanced dataset}
\usage{
DownSampleDataframe(input.data, name.of.feature.to.balance.on, seedVal)
}
\arguments{
\item{input.data}{A dataframe.}

\item{name.of.feature.to.balance.on}{Name of column in the input dataframe. Column has to be a factor. This column is the membership column that you want to balance for.}

\item{seedVal}{Numeric value to set seed for random number generator.}
}
\value{
List with two objects:
\enumerate{
\item ordered_balanced_data Dataframe with order of columns the same as the original data. Dataframe is a subset of the original dataframe because some observations were left out in order to balance the values in the selected column.
\item left_out_data Dataframe of the left out observations.
}
}
\description{
Down sample an imbalanced dataset to get a balanced dataset
}
