% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{generate.descriptive.plots}
\alias{generate.descriptive.plots}
\title{Use histograms and boxplots to get an general idea of what data looks like}
\usage{
generate.descriptive.plots(
  inputted.data,
  overall.plot.layout,
  plot.type,
  plot.features,
  plot.name
)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{overall.plot.layout}{A vector with two elements to specify how many rows and columns to include in the par plot. c(rows, cols)}

\item{plot.type}{A vector that specifies, for each subplot, should features be displayed in histogram or boxplot form. Elements can be "histogram" or "boxplot".}

\item{plot.features}{A vector that specifies, for each subplot, what features should be analyzed. The elements should be names of columns that contain continous numerical data.}

\item{plot.name}{A string for the name of the plot.}
}
\value{
No object is outputted, but a plot is displayed.
}
\description{
Creates a subplot for each specified feature. Each subplot can be histogram
or a boxplot.
}
\examples{

first.col <- c(10, 10, 30, 50, 20 , 10, 30)
second.col <- c(1, 2, 3, 1, 1, 1, 1)
input.data.frame <- as.data.frame(cbind(first.col, second.col))

generate.descriptive.plots(input.data.frame,
                          c(1, 2),
                          c("boxplot", "histogram"),
                          c("first.col", "second.col"),
                          "Example")


}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()}
}
\concept{Preprocessing functions}
