% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{GenerateElbowPlotPCA}
\alias{GenerateElbowPlotPCA}
\title{Create elbow plot to see how much total variance is explained by the components}
\usage{
GenerateElbowPlotPCA(inputted.data, column.to.do.PCA.on, scale.PCA)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{column.to.do.PCA.on}{A vector of strings that specify the column names that should be used for doing PCA.}

\item{scale.PCA}{Boolean to specify whether or not to scale columns before doing PCA.}
}
\value{
No object is returned. A plot will be created.
}
\description{
Create elbow plot to see how much total variance is explained by the components
}
