% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{GenerateExampleDataMachinelearnr}
\alias{GenerateExampleDataMachinelearnr}
\title{Produce example data set for demonstrating package functions}
\usage{
GenerateExampleDataMachinelearnr()
}
\value{
A dataframe.
}
\description{
Produce example data set for demonstrating package functions
}
