% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{GenerateParcoordForClusters}
\alias{GenerateParcoordForClusters}
\title{Generate parallel plot to show each observation and which cluster they belong in.}
\usage{
GenerateParcoordForClusters(
  inputted.data,
  cluster.assignment.column,
  x_axis_features
)
}
\arguments{
\item{inputted.data}{A dataframe where there are numerical columns for clustering.}

\item{cluster.assignment.column}{A string that specifies the name of the column in inputted.data that specify which cluster each observation in inputted.data belongs to.}

\item{x_axis_features}{A vector of strings that specify which features from inputted.data should be displayed on the x-axis of the parallel plot.}
}
\value{
No object is returned but a plot will be created.
}
\description{
Generate parallel plot to show each observation and which cluster they belong in.
}
\examples{

example.data <- data.frame(x = c(18, 21, 22, 24, 26, 26, 27, 30, 31,
                                 35, 39, 40, 41, 42, 44, 46, 47, 48, 49, 54, 35, 30),
                          y = c(10, 11, 22, 15, 12, 13, 14, 33, 39, 37, 44,
                                27, 29, 20, 28, 21, 30, 31, 23, 24, 40, 45))

#dev.new()
plot(example.data$x, example.data$y)

km.res <- stats::kmeans(example.data[,c("x", "y")], 3, nstart = 25, iter.max=10)

grouped <- km.res$cluster

example.data <- cbind(example.data, grouped)

GenerateParcoordForClusters(example.data, "grouped", c("x", "y"))


}
\seealso{
Other Clustering functions: 
\code{\link{CalcOptimalNumClustersForKMeans}()},
\code{\link{HierarchicalClustering}()},
\code{\link{generate.2D.clustering.with.labeled.subgroup}()},
\code{\link{generate.3D.clustering.with.labeled.subgroup}()},
\code{\link{generate.plots.comparing.clusters}()}
}
\concept{Clustering functions}
