% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{GeneratePC1andPC2PlotsWithAndWithoutOutliers}
\alias{GeneratePC1andPC2PlotsWithAndWithoutOutliers}
\title{Generate PC1 vs PC2 plots with and without outliers.}
\usage{
GeneratePC1andPC2PlotsWithAndWithoutOutliers(
  inputted.data,
  columns.to.do.PCA.on,
  scale.PCA,
  p.value.for.outliers
)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{columns.to.do.PCA.on}{A vector of strings that specify the column names that should be used for doing PCA.}

\item{scale.PCA}{Boolean to specify whether or not to scale columns before doing PCA.}

\item{p.value.for.outliers}{Outliers are defined as samples with either PC1 or PC2 values that
have a standard deviation value that meets a specified p-value
threshold.}
}
\value{
A List with two objects:
\enumerate{
\item Data after removing outliers for PC1 and PC2.
\item Data from outliers.
}

Plots will also be displayed.
}
\description{
Generate PC1 vs PC2 plots to visualize data with and without outliers.
also ouputs dataset with outliers removed.
}
\details{
Outliers are defined as samples with either PC1 or PC2 values that
have a standard deviation value that meets a specified p-value
threshold.
}
