% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Log2TargetDensityPlotComparison}
\alias{Log2TargetDensityPlotComparison}
\title{Do Log2 transformation on a column, and then compare with and without log2 transformation}
\usage{
Log2TargetDensityPlotComparison(inputted.data, col.name.to.log)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{col.name.to.log}{Name of column to do log transformation on. Column should be numerical continuous.}
}
\value{
No object is returned. A plot will be made.
}
\description{
Do Log2 transformation on a column, and then compare with and without log2 transformation
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
