% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{MultipleColumnsNormalCheckThenBoxCox}
\alias{MultipleColumnsNormalCheckThenBoxCox}
\title{Checks multiple columns in a dataframe to see if each is normally distributed. If not, then box-cox transform}
\usage{
MultipleColumnsNormalCheckThenBoxCox(
  input.data,
  names.of.dependent.variables,
  alpha.for.shapiro,
  output.lambda.in.col.name = TRUE
)
}
\arguments{
\item{input.data}{A dataframe.}

\item{names.of.dependent.variables}{Vector of strings where each element is the name of a column to assess for normality and potentially transform.}

\item{alpha.for.shapiro}{Numerical value from 0 to 1. Threshold for what is considered not normal.
If p-value is less than this threshold, then the data is considered not normal.}

\item{output.lambda.in.col.name}{Boolean indicating if the lambda used for boxcox should be included in the column name.}
}
\value{
A dataframe with the columns specified in names.of.dependent.variables.
}
\description{
Checks multiple columns in a dataframe to see if each is normally distributed. If not, then box-cox transform
}
