% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RecodeIdentifier}
\alias{RecodeIdentifier}
\title{Recode the identifier column of a dataset}
\usage{
RecodeIdentifier(input.data, name.of.identifier, file.name)
}
\arguments{
\item{input.data}{Dataframe with the data, including column to recode.}

\item{name.of.identifier}{String specifying the column name of the column with data to recode.}

\item{file.name}{Name of the file or absolute path of the file to write the output to.}
}
\value{
A modified dataframe with recoded identifier will be outputted. The recoded data will
also be written into a text file.
}
\description{
To protect patient confidentiality, sometimes the identifier in datasets
should be recoded.
}
