% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RemoveColWithAllZeros}
\alias{RemoveColWithAllZeros}
\title{Remove columns with all zeros}
\usage{
RemoveColWithAllZeros(inputted.data, columns.to.look.at = NULL)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{columns.to.look.at}{A vector of strings that specify the column names to look at. If no value is provided, then all columns are evaluated.}
}
\value{
A new dataframe with all-zero columns removed.
}
\description{
Remove columns with all zeros
}
