% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RemoveRowsBasedOnCol}
\alias{RemoveRowsBasedOnCol}
\title{Remove rows from the dataframe if the row contains a value in the specified columns}
\usage{
RemoveRowsBasedOnCol(inputted.data, columns.to.look.at, val.to.remove)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{columns.to.look.at}{A vector of strings with each string specifying the name of columns to look at.}

\item{val.to.remove}{Value to look for. If a row contains this value in the specified column(s), then the row is removed.}
}
\value{
Dataframe with some rows removed.
}
\description{
Remove rows from the dataframe if the row contains a value in the specified columns
}
