% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SplitIntoTrainTest}
\alias{SplitIntoTrainTest}
\title{Split into train and test}
\usage{
SplitIntoTrainTest(inputted.data, percent.to.train.with, seed)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{percent.to.train.with}{A number from 0 to 1 indicating percentage of data to be used for training.}

\item{seed}{A number to set the seed for random number generation.}
}
\value{
List with two objects. Train and test data, in that order.
}
\description{
Split into train and test
}
