% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SubsetDataByContinuousCol}
\alias{SubsetDataByContinuousCol}
\title{Subset data by two bounds on a continuous column}
\usage{
SubsetDataByContinuousCol(
  inputted.data,
  col.name.to.subset.on,
  lower.bound.percent,
  upper.bound.percent
)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{col.name.to.subset.on}{Name of column with continuous data for subsetting.}

\item{lower.bound.percent}{A numeric value from 0 to 1 that specifies the first bound.}

\item{upper.bound.percent}{A numeric value from 0 to 1 that specifies the second bound.}
}
\value{
A subsetted dataframe.
}
\description{
Example. If you want to only keep the bottom 25\% and top 75\% of data, then
the lower.bound.percent will be 0.25 and the upper.bound.percent will be 0.75.
}
