% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{TwoSampleTTest}
\alias{TwoSampleTTest}
\title{Performs two sample t-test on multiple features}
\usage{
TwoSampleTTest(
  input.data,
  column.names.to.use,
  name.of.class.column,
  file.name
)
}
\arguments{
\item{input.data}{A dataframe.}

\item{column.names.to.use}{Vector of strings where each string is the name of a column of numerical data to use for t-test.}

\item{name.of.class.column}{Name of column that contains labels for group1 and labels for group2. The labels can be any value.
For example, for group1 observations can have value of "1" for this column while group2 observations can have value of  "2" for this column.}

\item{file.name}{Name of the file to write output to.}
}
\value{
Dataframe where the rows are statistic values from the t-test and columns
are the features.
}
\description{
Performs two sample t-test on multiple features
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
