% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ZScoreChallengeOutliers}
\alias{ZScoreChallengeOutliers}
\title{Remove outliers based on Z score of a particular variable}
\usage{
ZScoreChallengeOutliers(inputted.data, column.to.perform.outlier.analysis)
}
\arguments{
\item{inputted.data}{A dataframe}

\item{column.to.perform.outlier.analysis}{Name of column in dataframe to evaluate for outliers. The column should contain continuous data.}
}
\value{
A dataframe with outlier rows removed.
}
\description{
If the Z-score of a sample for the selected column corresponds with a p-value
less than 0.05, then the sample is considered an outlier and removed.
}
\examples{


identifier.col <- c("a", "a", "a", "b", "b", "b", "c")
value.col <- c(1, 2, 3, 1, 1, 1, 100)
input.data.frame <- as.data.frame(cbind(identifier.col, value.col))

results <- ZScoreChallengeOutliers(input.data.frame, "value.col")

results

}
