% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{AddColBinnedToQuartiles}
\alias{AddColBinnedToQuartiles}
\title{Bin the values of a selected continuous column into 4 bins (quartiles) and add the bin assignments as a new column}
\usage{
AddColBinnedToQuartiles(inputted.data, col.name.to.bin, name.of.new.col)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{col.name.to.bin}{Name of a continuous column to use the values for binning.}

\item{name.of.new.col}{Name of the new column that will be added to the end of the dataframe. A suitable name would be "TARGET.quartile"}
}
\value{
A dataframe with bin assignment added as a factor column.
}
\description{
Bin the values of a selected continuous column into 4 bins (quartiles) and add the bin assignments as a new column
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
