% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{AddPCsToEnd}
\alias{AddPCsToEnd}
\title{Perform PCA}
\usage{
AddPCsToEnd(
  inputted.data,
  columns.to.do.PCA.on,
  scale.boolean,
  minimum.variance.percent
)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{columns.to.do.PCA.on}{A vector of strings that specify the column names to use for PCA.}

\item{scale.boolean}{Boolean specifying if data should be scaled before PCA.}

\item{minimum.variance.percent}{Value from 0-100 specifying how much percentage of total variance a principal component needs to meet or exceed to be kept.}
}
\value{
A List object that contains 3 elements:
\enumerate{
\item A dataframe with the PCs added to the end and the names of the PC columns.
\item A vector with the names of PCs.
\item The pca.results object from prcomp().
}
}
\description{
Perform PCA
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
