% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ConvertDataToPercentiles}
\alias{ConvertDataToPercentiles}
\title{Use percentiles to assess for outliers in multidimensional data}
\usage{
ConvertDataToPercentiles(input.data, upper_lower_bound_threshold)
}
\arguments{
\item{input.data}{A dataframe with the columns as continous values to be converted to percentile rank}

\item{upper_lower_bound_threshold}{A number from 0 to 1. The tails that should be considered as percentiles that are
too large or too small (upper_lower_bound_threshold, 1-upper_lower_bound_threshold)}
}
\value{
Original dataframe but values are converted to percentile and an additional column
is added to tally up how many features in each observation has an usually large or small percentile rank.
}
\description{
Takes a dataframe of values where the columns are continuous features
to be assess for outliers and rows are observations. For each feature, each observation
will receive a percentile rank for that feature. At the end, a column will be added
to the data that tallies how many features in a single observation has percentile
ranks in the top and bottom specified percentiles; this will be done for every observation.
Observations with many features in the top and bottom percentiles will be considered as
potential outliers. The tally column can be used to locate observations
that have many features with extreme values. As a result, the column
can be used to assess for potential outliers.
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
