% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{correlation.association.test}
\alias{correlation.association.test}
\title{Given two numerical data vector, determine the correlation}
\usage{
correlation.association.test(
  data1,
  data1.type,
  data1.name,
  data2,
  data2.type,
  data2.name
)
}
\arguments{
\item{data1}{Numerical data vector.}

\item{data1.type}{A string to specify what type of data is data1. Options are "continuous", "ordinal", "categorical".}

\item{data1.name}{String that specifies name to associate with data1.}

\item{data2}{Numerical data vector.}

\item{data2.type}{A string to specify what type of data is data1. Options are "continuous", "ordinal", "categorical".}

\item{data2.name}{String that specifies name to associate with data1.}
}
\value{
A vector of strings with two elements:
\enumerate{
\item Description of test performed
\item P-value
}
}
\description{
This is a work in progress function.
}
