% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{eval.classification.results}
\alias{eval.classification.results}
\title{Determine the performance of classification}
\usage{
eval.classification.results(actual_input, predicted_input, name)
}
\arguments{
\item{actual_input}{A vector that indicates the actual class for each observation.}

\item{predicted_input}{A vector that indicates the predicted class for each observation.}

\item{name}{A string that specifies the title to be used for plotting.}
}
\value{
No objects are created, but numbers and plots will be displayed.
}
\description{
The Matthew's Correlation Coefficient, proportion of total classifications
that are correct, and proportions of classifications correct for each
class are all displayed by this function. Plots are also displayed
to show performance of classification.
}
\examples{


id = c("1a", "1b", "1c", "1d", "1e", "1f", "1g", "2a", "2b", "2c", "2d", "2e", "2f", "3a",
       "3b", "3c", "3d", "3e", "3f", "3g", "3h", "3i")

x = c(18, 21, 22, 24, 26, 26, 27, 30, 31, 35, 39, 35, 30, 40, 41, 42, 44, 46, 47, 48, 49, 54)

y = c(10, 11, 22, 15, 12, 13, 14, 33, 39, 37, 44, 40, 45, 27, 29, 20, 28, 21, 30, 31, 23, 24)

a = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)

b = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)


actual = as.factor(c("1", "1", "1", "1", "1", "1", "1", "2", "2", "2", "2", "2", "2", "3", "3", "3",
       "3", "3", "3", "3", "3", "3"))

example.data <- data.frame(id, x, y, a, b, actual)

set.seed(2)
rf.result <- randomForest::randomForest(x=example.data[,c("x", "y", "a", "b")],
y=example.data[,"actual"], proximity=TRUE)

predicted <- rf.result$predicted
actual <- example.data[,"actual"]

eval.classification.results(actual, predicted, "Example")


}
\seealso{
Other Classification functions: 
\code{\link{LOOCVPredictionsRandomForestAutomaticMtryAndNtree}()},
\code{\link{LOOCVRandomForestClassificationMatrixForPheatmap}()},
\code{\link{RandomForestAutomaticMtryAndNtree}()},
\code{\link{RandomForestClassificationGiniMatrixForPheatmap}()},
\code{\link{RandomForestClassificationPercentileMatrixForPheatmap}()},
\code{\link{find.best.number.of.trees}()}
}
\concept{Classification functions}
