% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{GenerateExampleDataMachinelearnr}
\alias{GenerateExampleDataMachinelearnr}
\title{Produce example data set for demonstrating package functions}
\usage{
GenerateExampleDataMachinelearnr()
}
\value{
A dataframe.
}
\description{
Produce example data set for demonstrating package functions
}
\seealso{
Other Classification functions: 
\code{\link{CVPredictionsRandomForest}()},
\code{\link{CVRandomForestClassificationMatrixForPheatmap}()},
\code{\link{LOOCVPredictionsRandomForestAutomaticMtryAndNtree}()},
\code{\link{LOOCVRandomForestClassificationMatrixForPheatmap}()},
\code{\link{RandomForestAutomaticMtryAndNtree}()},
\code{\link{RandomForestClassificationGiniMatrixForPheatmap}()},
\code{\link{RandomForestClassificationPercentileMatrixForPheatmap}()},
\code{\link{eval.classification.results}()},
\code{\link{find.best.number.of.trees}()}
}
\concept{Classification functions}
