% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Log2TargetDensityPlotComparison}
\alias{Log2TargetDensityPlotComparison}
\title{Do Log2 transformation on a column, and then compare with and without log2 transformation}
\usage{
Log2TargetDensityPlotComparison(inputted.data, col.name.to.log)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{col.name.to.log}{Name of column to do log transformation on. Column should be numerical continuous.}
}
\value{
No object is returned. A plot will be made.
}
\description{
Do Log2 transformation on a column, and then compare with and without log2 transformation
}
