% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LookAtPCFeatureLoadings}
\alias{LookAtPCFeatureLoadings}
\title{Principal component feature loadings}
\usage{
LookAtPCFeatureLoadings(pca.results, pc.to.look.at)
}
\arguments{
\item{pca.results}{Object outputted from stats::prcomp()}

\item{pc.to.look.at}{Numeric value indicating which principal component to look at.}
}
\value{
No objects are outputted but a plot is created.
}
\description{
Principal component feature loadings
}
