% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{NormalCheckThenBoxCoxTransform}
\alias{NormalCheckThenBoxCoxTransform}
\title{Checks if the data is normally distributed using Shapiro test. If not normal, then boxcox transform.}
\usage{
NormalCheckThenBoxCoxTransform(input.data, alpha.for.shapiro)
}
\arguments{
\item{input.data}{A numerical data vector.}

\item{alpha.for.shapiro}{Numerical value from 0 to 1. Threshold for what is considered not normal.
If p-value is less than this threshold, then the data is considered not normal.}
}
\value{
A List with 4 elements:
\enumerate{
\item If data is non-normal, then a vector of the transformed data is outputted. If data is normal, then this is NULL.
\item If data is non-normal, then a number specifying the lambda used for boxcox is outputted. If data is normal, then this is NULL.
\item P-value from the Shapiro test.
\item Boolean indicating if boxcox transformation was performed.
}
}
\description{
Cannot use boxcox if data has zeroes. If data has zeros, then do this:
Add 1 if most values are greater than 1, else  if most of values <1,
multiply 10 or 100, then add 1. Then do boxcox.
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
