% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RecodeIdentifier}
\alias{RecodeIdentifier}
\title{Recode the identifier column of a dataset}
\usage{
RecodeIdentifier(input.data, name.of.identifier, file.name)
}
\arguments{
\item{input.data}{Dataframe with the data, including column to recode.}

\item{name.of.identifier}{String specifying the column name of the column with data to recode.}

\item{file.name}{Name of the file or absolute path of the file to write the output to.}
}
\value{
A modified dataframe with recoded identifier will be outputted. The recoded data will
also be written into a text file.
}
\description{
To protect patient confidentiality, sometimes the identifier in datasets
should be recoded.
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
