% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SplitIntoTrainTest}
\alias{SplitIntoTrainTest}
\title{Split into train and test}
\usage{
SplitIntoTrainTest(inputted.data, percent.to.train.with, seed)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{percent.to.train.with}{A number from 0 to 1 indicating percentage of data to be used for training.}

\item{seed}{A number to set the seed for random number generation.}
}
\value{
List with two objects. Train and test data, in that order.
}
\description{
Split into train and test
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
