% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{StabilityTestingAcrossVisits}
\alias{StabilityTestingAcrossVisits}
\title{Assess stability of values that correspond to a single identifier}
\usage{
StabilityTestingAcrossVisits(
  inputted.data,
  col.name.of.unique.identifier,
  value.to.evaluate
)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{col.name.of.unique.identifier}{A string that specifies name of column in inputted.data containing unique identifiers.}

\item{value.to.evaluate}{A string that specifies name of column in inputted.data to look at for stability of values.}
}
\value{
A dataframe with the unique.identifiers as rows and the standard deviation of the values.
If a individual only has one row in the original dataframe, the SD value will be NA.
}
\description{
Assess stability of values that correspond to a single identifier
}
\examples{
identifier.col <- c("a", "a", "a", "b", "b", "b", "c")
value.col <- c(1, 2, 3, 1, 1, 1, 5)
input.data.frame <- as.data.frame(cbind(identifier.col, value.col))

results <- StabilityTestingAcrossVisits(input.data.frame, "identifier.col", "value.col")

results




}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{SubsetDataByContinuousCol}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
