% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SubsetDataByContinuousCol}
\alias{SubsetDataByContinuousCol}
\title{Subset data by two bounds on a continuous column}
\usage{
SubsetDataByContinuousCol(
  inputted.data,
  col.name.to.subset.on,
  lower.bound.percent,
  upper.bound.percent
)
}
\arguments{
\item{inputted.data}{A dataframe.}

\item{col.name.to.subset.on}{Name of column with continuous data for subsetting.}

\item{lower.bound.percent}{A numeric value from 0 to 1 that specifies the first bound.}

\item{upper.bound.percent}{A numeric value from 0 to 1 that specifies the second bound.}
}
\value{
A subsetted dataframe.
}
\description{
Example. If you want to only keep the bottom 25\% and top 75\% of data, then
the lower.bound.percent will be 0.25 and the upper.bound.percent will be 0.75.
}
\seealso{
Other Preprocessing functions: 
\code{\link{AddColBinnedToBinary}()},
\code{\link{AddColBinnedToQuartiles}()},
\code{\link{AddPCsToEnd}()},
\code{\link{ConvertDataToPercentiles}()},
\code{\link{CorAssoTestMultipleWithErrorHandling}()},
\code{\link{DownSampleDataframe}()},
\code{\link{GenerateElbowPlotPCA}()},
\code{\link{GeneratePC1andPC2PlotsWithAndWithoutOutliers}()},
\code{\link{Log2TargetDensityPlotComparison}()},
\code{\link{LookAtPCFeatureLoadings}()},
\code{\link{MultipleColumnsNormalCheckThenBoxCox}()},
\code{\link{NormalCheckThenBoxCoxTransform}()},
\code{\link{RanomlySelectOneRowForEach}()},
\code{\link{RecodeIdentifier}()},
\code{\link{RemoveColWithAllZeros}()},
\code{\link{RemoveRowsBasedOnCol}()},
\code{\link{RemoveSamplesWithInstability}()},
\code{\link{SplitIntoTrainTest}()},
\code{\link{StabilityTestingAcrossVisits}()},
\code{\link{TwoSampleTTest}()},
\code{\link{ZScoreChallengeOutliers}()},
\code{\link{captureSessionInfo}()},
\code{\link{correlation.association.test}()},
\code{\link{describeNumericalColumnsWithLevels}()},
\code{\link{describeNumericalColumns}()},
\code{\link{generate.descriptive.plots.save.pdf}()},
\code{\link{generate.descriptive.plots}()}
}
\concept{Preprocessing functions}
